final fix for warning "array subscript has type char" we are now using the same syntax as strtolower() [oscam-string.c] see thread: http://www.streamboard.tv/wbb2/thread.php?postid=535127
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@10379 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-webif-tpl.c
@@ -723,7 +723,7 @@ char *urlencode(struct templatevars *vars, char *str)
 	char *pstr = str, *pbuf = buf;
 	while(*pstr)
 	{
-		if(isalnum(*pstr) || *pstr == '-' || *pstr == '_' || *pstr == '.' || *pstr == '~') { *pbuf++ = *pstr; }
+		if(isalnum((uchar)*pstr) || *pstr == '-' || *pstr == '_' || *pstr == '.' || *pstr == '~') { *pbuf++ = *pstr; }
 		else if(*pstr == ' ') { *pbuf++ = '+'; }
 		else
 		{
